requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
botocore==1.34.149
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
outcome==1.3.0.post0
packaging==23.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.2
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
webdriver-manager==4.0.1
wsproto==1.2.0